requirements.txt
absl-py==2.3.1
addict==2.4.0
antlr4-python3-runtime==4.9.3
antspyx==0.6.1
asttokens==3.0.1
attrs==25.4.0
backports.zstd @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_backports.zstd_1767044984/work
blinker==1.9.0
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1764016952863/work
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1767500808759/work/certifi
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1760437218288/work
click==8.3.1
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1733218098505/work
comm==0.2.3
ConfigArgParse==1.7.1
contourpy==1.3.2
cycler==0.12.1
dash==3.4.0
dataclasses @ file:///home/conda/feedstock_root/build_artifacts/dataclasses_1628958434797/work
decorator==5.2.1
exceptiongroup==1.3.1
executing==2.2.1
fastjsonschema==2.21.2
filelock @ file:///home/conda/feedstock_root/build_artifacts/filelock_1768022204021/work
Flask==3.1.2
fonttools==4.61.1
fsspec==2026.1.0
fvcore @ file:///home/conda/feedstock_root/build_artifacts/fvcore_1671623667463/work
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1762946744175/work
grpcio==1.76.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hydra-core==1.3.2
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1760286409563/work
ImageIO==2.37.2
importlib_metadata==8.7.1
iopath @ file:///home/conda/feedstock_root/build_artifacts/iopath_1723846728822/work
ipython==8.38.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jinja2_1764517220/work
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_core==5.9.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lazy_loader==0.4
lxml==6.0.2
Markdown==3.10.1
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1759055168201/work
matplotlib==3.7.5
matplotlib-inline==0.2.1
monai==1.3.2
mpmath @ file:///home/conda/feedstock_root/build_artifacts/mpmath_1733302684489/work
narwhals==2.15.0
nbformat==5.5.0
nest-asyncio==1.6.0
networkx @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_networkx_1731521053/work
nibabel==5.2.0
nilearn==0.10.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1682210193659/work
omegaconf==2.3.0
open3d==0.16.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1769093650/work
pandas==2.0.3
parso==0.8.5
patsy==1.0.2
pexpect==4.9.0
Pillow @ file:///home/conda/feedstock_root/build_artifacts/pillow_1696149248731/work
platformdirs==4.5.1
plotly==6.5.2
pooch==1.8.2
portalocker @ file:///home/conda/feedstock_root/build_artifacts/portalocker_1757395643892/work
prompt_toolkit==3.0.52
protobuf==6.33.4
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
pyquaternion==0.9.9
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil==2.9.0.post0
pytorch-ranger==0.1.1
pytorch3d==0.7.8
pytz==2025.2
pyvista==0.44.1
PyWavelets==1.8.0
PyYAML @ file:///home/conda/feedstock_root/build_artifacts/pyyaml_1758891829620/work
referencing==0.37.0
requests @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_requests_1766926974/work
retrying==1.4.2
rpds-py==0.30.0
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
scooby==0.11.0
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.6
sympy @ file:///home/conda/feedstock_root/build_artifacts/sympy_1745946051654/work
tabulate @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_tabulate_1765140047/work
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1767096303160/work
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.1.0
torch-optimizer==0.3.0
torchvision==0.16.0
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1735661334605/work
traitlets==5.14.3
trimesh==4.1.3
triton==2.1.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_typing_extensions_1756220668/work
tzdata==2025.3
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1767817748113/work
vtk==9.5.2
wcwidth==0.5.0
webcolors==25.10.0
Werkzeug==3.1.5
widgetsnbextension==4.0.15
yacs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_yacs_1737043264/work
zipp==3.23.0
